# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: goimg-datalayer API
  version: 2.0.0
  description: |
    A self-hosted image gallery backend inspired by Flickr and Chevereto.

    Features:
    - User management with JWT authentication
    - Image upload with automatic variant generation
    - Albums, tags, and social features (likes, comments)
    - Content moderation and abuse reporting
    - RESTful API with RFC 7807 error responses

    **Phase 2 Complete**: All Phase 2 features (Sprints 10-15) implemented.
    **Phase 3 Progress**: Sprints 16-20 complete (oEmbed, nested albums, custom variants, test coverage, trending tags, featured picks, groups/communities).
    **Current Sprint**: Sprint 23 in progress (test coverage improvement, documentation updates).
    Future features in backlog: video support (Sprint 21) and account tiers (Sprint 22).

  contact:
    name: goimg-datalayer API Team
    url: https://github.com/yegamble/goimg-datalayer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.goimg.com/v1
    description: Production server (example)
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      description: |
        Create a new user account with email, username, and password.

        Validation:
        - Email must be unique and not from disposable provider
        - Username must be unique, 3-32 chars, alphanumeric
        - Password must be 12-128 chars with complexity requirements
        - Password is checked against Have I Been Pwned (HIBP) database using k-anonymity
          (only first 5 chars of SHA-1 hash are sent, preserving privacy)

        Security:
        - Compromised passwords (found in data breaches) are rejected
        - HIBP check uses k-anonymity to protect user privacy
        - If HIBP service is unavailable, registration proceeds (fail-open)
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                  example: user@example.com
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                  pattern: '^[a-zA-Z0-9_-]+$'
                  example: johndoe
                password:
                  type: string
                  format: password
                  minLength: 12
                  maxLength: 128
                  example: MySecureP@ssw0rd123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                  username:
                    type: string
                  created_at:
                    type: string
                    format: date-time
        '400':
          description: |
            Validation error. Possible error types:
            - `validation-error`: Field validation failed (e.g., password too short)
            - `password-compromised`: Password found in a data breach (HIBP check)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              examples:
                validation_error:
                  summary: Field validation error
                  value:
                    type: https://api.goimg.com/problems/validation-error
                    title: Validation Error
                    status: 400
                    detail: Request validation failed
                    traceId: abc123-def456
                    errors:
                      - field: password
                        message: password must be at least 12 characters
                password_compromised:
                  summary: Password found in data breach
                  value:
                    type: https://api.goimg.com/problems/password-compromised
                    title: Password Compromised
                    status: 400
                    detail: This password has been found in a data breach and cannot be used. Please choose a different password.
                    traceId: abc123-def456
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: https://api.goimg.com/problems/conflict
                title: Conflict
                status: 409
                detail: Email or username already exists
                traceId: abc123-def456
  /auth/login:
    post:
      tags: [auth]
      summary: User login
      description: |
        Authenticate with email and password to receive JWT tokens.

        Security:
        - Rate limit: 5 attempts per minute per IP
        - Account lockout after 5 failed attempts (15 min)
        - Generic error messages to prevent account enumeration
        - Random response delay (100-300ms) to prevent timing attacks
        - Constant-time password comparison (Argon2id)
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: MySecureP@ssw0rd123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: https://api.goimg.com/problems/unauthorized
                title: Unauthorized
                status: 401
                detail: Invalid email or password
                traceId: abc123-def456
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      description: |
        Exchange a refresh token for a new access token and refresh token.

        Security:
        - Refresh token rotation (old token invalidated)
        - Replay attack detection (token reuse revokes entire family)
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/logout:
    post:
      tags: [auth]
      summary: User logout
      description: Invalidate the current refresh token
      operationId: logoutUser
      responses:
        '204':
          description: Logout successful
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request password reset
      description: |
        Send a password reset email to the user. Always returns 200 regardless
        of whether the email exists (anti-enumeration protection).
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Password reset email sent (or silently ignored if email not found)
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset password
      description: |
        Reset the user's password using a valid reset token received by email.
        On success all existing sessions are revoked for security.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                  description: Password reset token from the email link
                new_password:
                  type: string
                  format: password
                  minLength: 12
                  maxLength: 128
                  example: NewSecurePassword123!
      responses:
        '204':
          description: Password reset successfully
        '400':
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ==================== Two-Factor Authentication ====================
...
  /auth/guest:
    post:
      tags: [auth]
      summary: Create guest session
      description: |
        Create a temporary guest session for anonymous uploads.

        Guest sessions allow users to upload images without registering.
        The guest account expires after 30 days, and uploads can be
        claimed by registering or linking to an existing account.

        Rate limits:
        - 10 guest sessions per hour per IP address

        The returned token can be used for subsequent API calls.
      operationId: createGuestSession
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                user_agent:
                  type: string
                  description: Optional browser/client user agent
                  maxLength: 500
      responses:
        '201':
          description: Guest session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '429':
          description: Rate limit exceeded for guest session creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer authentication. Include the access token in the Authorization header:

        ```
        Authorization: Bearer <access_token>
        ```

        Access tokens expire after 15 minutes. Use the refresh endpoint to obtain a new token.
...
  schemas:
...
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: https://api.goimg.com/problems/validation-error
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: Validation Error
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: Request validation failed
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
          example: /api/v1/images
        traceId:
          type: string
          description: A unique identifier for tracing the request
          example: abc123-def456
        errors:
          type: array
          description: Detailed validation errors (for 400 responses)
          items:
            type: object
            properties:
              field:
                type: string
                example: title
              message:
                type: string
                example: must not exceed 255 characters

    # ==================== Session ====================
...